heredoc

+++ execution/heredoc.c
@@ -64,6 +64,7 @@ int	heredoc(t_executor *executor, t_data *data)
 {
 	char	*temp;
 	int		i;
+	int		fd;
 	char	*delimiter;
 
 	i = 0;
@@ -72,10 +73,14 @@ int	heredoc(t_executor *executor, t_data *data)
 	delimiter = find_delimiter(delimiter, data, i, temp);
 	if (!delimiter)
 		return (0);
-	executor->in = open(".heredoc_temp", O_CREAT | O_WRONLY | O_TRUNC, 0644);
+	fd = open(".heredoc_temp", O_CREAT | O_WRONLY | O_TRUNC, 0644);
+	if (fd < 0)
+		return (free(delimiter), 0);
+	heredoc_loop(delimiter, fd);
+	close(fd);
+	executor->in = open(".heredoc_temp", O_RDONLY);
 	if (executor->in < 0)
 		return (free(delimiter), 0);
-	heredoc_loop(delimiter, executor->in);
 	if (unlink(".heredoc_temp") == -1)
 		return (perror("unlink"), 1);
 	return (free(delimiter), 1);
